==> code/07.construct_phyloseq.R <==
# construct phyloseq post KTU computation
library(phyloseq)
library(tidyverse)

# copy ASV DNA seq to refseq slot and simplify otu_table names
source("code/functions/taxanames2refseq_ps.r")
# ktu object (list) to input for idTaxa (matrix)
source("code/functions/klustering2otutable.r")

# load seqs with taxonomy
load("data/intermediate/taxid.Rdata")
# load OTU tables
ktu16s <-
  readRDS("data/intermediate/ktu_bact.rds") %>%
  klustering2otutable()
ktuits <-
  readRDS("data/intermediate/ktu_fung.rds") %>%
  klustering2otutable()
# meta
meta <- readRDS("data/intermediate/meta.rds")

# assert all samples in the OTU table are present in metadata
assertthat::assert_that(
  all(colnames(ktu16s) %in% rownames(meta)))
assertthat::assert_that(
  all(colnames(ktuits) %in% rownames(meta)))

# construct phyloseq object
ps_16sktu <-
  phyloseq::phyloseq(
    t(otu_table(ktu16s, taxa_are_rows = TRUE)),
    sample_data(meta),
    tax_table(taxid_16S)) %>%
    taxanames2refseq_ps()

ps_itsktu <-
  phyloseq::phyloseq(
    t(otu_table(ktuits, taxa_are_rows = TRUE)),
    sample_data(meta),
    tax_table(taxid_its)) %>%
    taxanames2refseq_ps()

#save phyloseq objects
save(ps_16sktu, ps_itsktu, file = "data/intermediate/psKTU.Rdata")
